pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen>=0.7.4",
    "faiss-cpu>=1.10.0",
    "fastapi[standard]>=0.115.8",
    "google-search-results>=2.4.2",
    "groq>=0.18.0",
    "httpx>=0.28.1",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-core>=0.3.37",
    "langchain-google-genai>=2.0.9",
    "langchain-groq>=0.2.4",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.6",
    "langchainhub>=0.1.21",
    "litellm>=1.61.11",
    "pypdf>=5.3.0",
    "python-dotenv>=1.0.1",
    "scholarly>=1.7.11",
    "sentence-transformers>=3.4.1",
    "streamlit>=1.42.1",
    "tiktoken>=0.9.0",
]

[project.scripts]
research-agent = "research_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
